refactor: validate a list of markings
